Relocate SPI to modern package name
