+++ src/main/java/io/vertx/jdbcclient/impl/actions/JDBCPreparedQuery.java
@@ -18,7 +18,7 @@ package io.vertx.jdbcclient.impl.actions;
 
 import io.vertx.core.buffer.Buffer;
 import io.vertx.core.json.JsonArray;
-import io.vertx.ext.jdbc.spi.JDBCColumnDescriptorProvider;
+import io.vertx.jdbcclient.spi.JDBCColumnDescriptorProvider;
 import io.vertx.ext.sql.SQLOptions;
 import io.vertx.jdbcclient.SqlOutParam;
 import io.vertx.sqlclient.Row;
