+++ src/main/java/io/vertx/jdbcclient/impl/actions/JDBCPreparedStatement.java
@@ -15,7 +15,7 @@
  */
 package io.vertx.jdbcclient.impl.actions;
 
-import io.vertx.ext.jdbc.spi.JDBCColumnDescriptorProvider;
+import io.vertx.jdbcclient.spi.JDBCColumnDescriptorProvider;
 import io.vertx.sqlclient.impl.ParamDesc;
 import io.vertx.sqlclient.impl.PreparedStatement;
 import io.vertx.sqlclient.impl.RowDesc;
@@ -23,8 +23,6 @@ import io.vertx.sqlclient.impl.TupleInternal;
 
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
 
 public class JDBCPreparedStatement implements PreparedStatement {
