+++ src/main/java/io/vertx/jdbcclient/impl/actions/JDBCQueryAction.java
@@ -11,8 +11,8 @@
 
 package io.vertx.jdbcclient.impl.actions;
 
-import io.vertx.ext.jdbc.spi.JDBCColumnDescriptorProvider;
-import io.vertx.ext.jdbc.spi.JDBCDecoder;
+import io.vertx.jdbcclient.spi.JDBCColumnDescriptorProvider;
+import io.vertx.jdbcclient.spi.JDBCDecoder;
 import io.vertx.ext.sql.SQLOptions;
 import io.vertx.jdbcclient.impl.JDBCRow;
 import io.vertx.sqlclient.Row;
