+++ src/main/java/io/vertx/jdbcclient/spi/JDBCEncoderImpl.java
@@ -14,14 +14,12 @@
  * You may elect to redistribute this code under either of these licenses.
  */
 
-package io.vertx.ext.jdbc.spi.impl;
+package io.vertx.jdbcclient.spi;
 
 import io.vertx.core.impl.logging.Logger;
 import io.vertx.core.impl.logging.LoggerFactory;
 import io.vertx.core.json.JsonArray;
 import io.vertx.jdbcclient.impl.actions.JDBCStatementHelper;
-import io.vertx.ext.jdbc.spi.JDBCColumnDescriptorProvider;
-import io.vertx.ext.jdbc.spi.JDBCEncoder;
 import io.vertx.jdbcclient.impl.actions.JDBCColumnDescriptor;
 import io.vertx.sqlclient.Tuple;
 
@@ -51,7 +49,7 @@ public class JDBCEncoderImpl implements JDBCEncoder {
     return doEncode(provider.apply(pos), input.getValue(pos - 1));
   }
 
-  protected Object doEncode(JDBCColumnDescriptor descriptor, Object javaValue) {
+  public Object doEncode(JDBCColumnDescriptor descriptor, Object javaValue) {
     if (javaValue == null) {
       return null;
     }
