+++ src/test/java/io/vertx/it/MSSQLDecoder.java
@@ -17,7 +17,7 @@
 package io.vertx.it;
 
 import io.vertx.jdbcclient.impl.actions.SQLValueProvider;
-import io.vertx.ext.jdbc.spi.impl.JDBCDecoderImpl;
+import io.vertx.jdbcclient.spi.JDBCDecoderImpl;
 import io.vertx.jdbcclient.impl.actions.JDBCColumnDescriptor;
 import microsoft.sql.DateTimeOffset;
 import microsoft.sql.Types;
