+++ src/test/java/io/vertx/jdbcclient/JDBCPoolColumnDescOracleExtTest.java
@@ -14,7 +14,7 @@ package io.vertx.jdbcclient;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
 import io.vertx.jdbcclient.impl.actions.SQLValueProvider;
-import io.vertx.ext.jdbc.spi.impl.JDBCDecoderImpl;
+import io.vertx.jdbcclient.spi.JDBCDecoderImpl;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 import io.vertx.jdbcclient.impl.actions.JDBCColumnDescriptor;
